fix: #164 virtio-net + UEFI crash
When using virtio-net with an UEFI firmware (bootrom),
xhyve crashes on an assertion. See the related bug
#164

The complete details are not clear to me, but it seems
that during boot, the firmware reads on the interface but
doesn't provide big enough read buffers. It seems that
the space for VLAN headers is missing.

This patch drops the related reads, so normal booting can
occur. Network boot probably doesn't work, though.

Signed-off-by: Antoine Martin amartin@unowhy.com